Renaming the namespace Microsoft.Framework.WebEncoders to System.Text.Encodings.Web.Tests
- Namespace Microsoft.Framework.WebEncoders renamed to System.Text.Encodings.Web.Tests
