+++ src/System.Text.Encodings.Web/tests/AllowedCharsBitmapTests.cs
@@ -8,7 +8,7 @@ using System.Text.Internal;
 using System.Text.Unicode;
 using Xunit;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     public class AllowedCharsBitmapTests
     {
