+++ src/System.Text.Encodings.Web/tests/ConfigurableScalarTextEncoder.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Text.Encodings.Web;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     /// <summary>
     /// Dummy encoder used for unit testing.
