+++ src/System.Text.Encodings.Web/tests/EncoderCommonTests.cs
@@ -6,7 +6,7 @@ using System;
 using System.Text.Encodings.Web;
 using Xunit;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     public class EncoderCommonTests
     {
