+++ src/System.Text.Encodings.Web/tests/HtmlEncoderTests.cs
@@ -9,7 +9,7 @@ using System.Text.Encodings.Web;
 using System.Text.Unicode;
 using Xunit;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     public class HtmlEncoderTests
     {
