+++ src/System.Text.Encodings.Web/tests/IJavaScriptStringEncoder.cs
@@ -4,7 +4,7 @@
 
 using System.IO;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     /// <summary>
     /// Provides services for JavaScript-escaping strings.
