+++ src/System.Text.Encodings.Web/tests/JavaScriptStringEncoderTests.cs
@@ -11,7 +11,7 @@ using System.Text.Encodings.Web;
 using System.Text.Unicode;
 using Xunit;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     public partial class JavaScriptStringEncoderTests
     {
