+++ src/System.Text.Encodings.Web/tests/PerformanceTests.cs
@@ -11,7 +11,7 @@ using Xunit;
 using Xunit.Abstractions;
 
 #if RELEASE
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     public class PerformanceTests
     {
