+++ src/System.Text.Encodings.Web/tests/ScalarTestEncoder.cs
@@ -8,7 +8,7 @@ using System.IO;
 using System.Runtime.CompilerServices;
 using System.Text.Encodings.Web;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     /// <summary>
     /// Dummy encoder used for unit testing.
