+++ src/System.Text.Encodings.Web/tests/TemporaryEncoderExtensions.cs
@@ -5,7 +5,7 @@
 using System;
 using System.IO;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     /// <summary>
     /// Helpful extension methods for the encoder classes.
