+++ src/System.Text.Encodings.Web/tests/TextEncoderTests.cs
@@ -10,7 +10,7 @@ using System.Text;
 using System.Text.Encodings.Web;
 using Xunit;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     public class TextEncoderTests
     {
