+++ src/System.Text.Encodings.Web/tests/UnicodeEncoderBaseTests.cs
@@ -10,7 +10,7 @@ using System.Text.Encodings.Web;
 using System.Text.Unicode;
 using Xunit;
 
-namespace Microsoft.Framework.WebEncoders
+namespace System.Text.Encodings.Web.Tests
 {
     public class UnicodeEncoderBaseTests
     {
